Fixed a very pesky memory bug
This bug was a complete emotional rollercoaster to find. Because ... this Assume init should never run (thus never causing ub), however for some reason, only on android, does this code run, ... run into the ground.
Just replaced this with a option, same size, lot more explicit (and it doesn't crash :) )